backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
anthropic==0.39.0
python-dotenv==1.0.1
pydantic==2.9.2
python-multipart==0.0.12
aiofiles==24.1.0
httpx==0.27.2
apscheduler==3.10.4
